+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- Columns for Quick Links! (finally, I know)
+
 ## [1.15.0] - 2018-08-03
 ### Added
 - Translations for Czech, Dutch and Traditional Chinese!